// src/app/models/time.model.ts
export interface DateTime {
    year: number
    month: number
    day: number
    hour: number
    minute: number
    seconds: number
    milliSeconds: number
    dateTime: string
    date: string
    time: string
    timeZone: string
    dayOfWeek: string
    dstActive: boolean
  }
  
  export interface DateFormat {
    year: string,
    month: string,
    day: string
  }